requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
drake==1.42.0
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.58.5
gradescope-utils==0.5.0
greenlet==3.2.4
ipykernel==6.29.5
ipympl==0.9.7
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lxml==6.0.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
Mosek==11.0.25
mpld3==0.5.10
narwhals==1.48.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydot==4.0.1
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
rpds-py==0.26.0
scipy==1.16.0
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
timeout-decorator==0.5.0
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
underactuated==2025.2.16
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.14
